Incorperated the RBModel class.
The RBModel class is now incorperated into the project. It is created
RBWindow's constructor and passed to all the listeners.

There are several issues with the RBModel class. First the class does
not have a refrence to the window so cannot tell it to redraw. Second
the class does not seam to have any knowlege of the side pane's buttons
which have the root node's options. Third the click function seams
clunky in that everything is determined by one parameter so a button's
index has to be changed to be used.