Finished adding window functionality.
The main pane in the window can now switch between a pane with six
buttons and a pane with a text area and an image. This can be done
with the functions showButtonPane() and showTextPane() respectively.

The contents of the text pane can be changed with the function
refreshTextPane(String text, String image).